camera + refactoring
