+++ hardware_interfaces/sensors/webots_imu.py
@@ -1,7 +1,7 @@
 # TODO: add docs
 # TODO: add config read
 import logging
-from imu_sensor_interface import IMUSensorInterface
+from .imu_sensor_interface import IMUSensorInterface
 from controller import InertialUnit
 
 
@@ -24,16 +24,16 @@ class WebotsIMU(IMUSensorInterface):
     def start(self):
         try:
             self.__in_unit.enable(100)
-            self.__status = 'enabled'
+            self.status = 'enabled'
         except Exception as start_exception:
-            self.__status = 'disabled'
+            self.status = 'disabled'
             logging.error(start_exception)
 
-        self.__status = 'enabled'
+        self.status = 'enabled'
 
     def _get_orientation(self):
         orientation = None
-        if self.__status == 'enabled':
+        if self.status == 'enabled':
             orientation = self.__in_unit.getRollPitchYaw()
         else:
             logging.error("IMU is not started")
