+++ main_scrpit.py
@@ -5,8 +5,8 @@ import time
 def main():
     #rclpy.init()
     kondo_webots_cm = ControllerManager(type='Webots', robot=Robot())
-    kondo_webots_cm.controllers['servos'].set_position('head_yaw', -0.5)
-    kondo_webots_cm.controllers['servos'].set_position('head_pitch', 1.0)
+    #kondo_webots_cm.controllers['servos'].set_position('head_yaw', -0.5)
+    #kondo_webots_cm.controllers['servos'].set_position('head_pitch', 1.0)
 
     kondo_webots_cm.start()
